// src/http/generated/lunajoy-server.schemas.ts
export type CreateDailyLogBodyMood = typeof CreateDailyLogBodyMood[keyof typeof CreateDailyLogBodyMood];


export const CreateDailyLogBodyMood = {
  VERY_SAD: 'VERY_SAD',
  SAD: 'SAD',
  NEUTRAL: 'NEUTRAL',
  HAPPY: 'HAPPY',
  VERY_HAPPY: 'VERY_HAPPY',
} as const;

export type CreateDailyLogBodyAnxietyLevel = typeof CreateDailyLogBodyAnxietyLevel[keyof typeof CreateDailyLogBodyAnxietyLevel];


export const CreateDailyLogBodyAnxietyLevel = {
  CALM: 'CALM',
  SLIGHTLY_ANXIOUS: 'SLIGHTLY_ANXIOUS',
  MODERATE: 'MODERATE',
  HIGH: 'HIGH',
  OVERWHELMED: 'OVERWHELMED',
} as const;

export type CreateDailyLogBodyPhysicalActivityItemActivityType = typeof CreateDailyLogBodyPhysicalActivityItemActivityType[keyof typeof CreateDailyLogBodyPhysicalActivityItemActivityType];


export const CreateDailyLogBodyPhysicalActivityItemActivityType = {
  WALKING: 'WALKING',
  RUNNING: 'RUNNING',
  CYCLING: 'CYCLING',
  SWIMMING: 'SWIMMING',
  YOGA: 'YOGA',
  OTHER: 'OTHER',
} as const;

export type CreateDailyLogBodyPhysicalActivityItemIntensity = typeof CreateDailyLogBodyPhysicalActivityItemIntensity[keyof typeof CreateDailyLogBodyPhysicalActivityItemIntensity];


export const CreateDailyLogBodyPhysicalActivityItemIntensity = {
  LOW: 'LOW',
  MODERATE: 'MODERATE',
  HIGH: 'HIGH',
} as const;
// ...
export type CreateDailyLogBodySleepSleepQualityRating = typeof CreateDailyLogBodySleepSleepQualityRating[keyof typeof CreateDailyLogBodySleepSleepQualityRating];


export const CreateDailyLogBodySleepSleepQualityRating = {
  VERY_POOR: 'VERY_POOR',
  POOR: 'POOR',
  AVERAGE: 'AVERAGE',
  GOOD: 'GOOD',
  EXCELLENT: 'EXCELLENT',
} as const;

export type CreateDailyLogBodySleep = {
  /**
     * @minimum 0
     * @maximum 24
     */
  sleepDurationInHours: number;
  sleepQualityRating: CreateDailyLogBodySleepSleepQualityRating;
  disturbances?: string;
};

export type CreateDailyLogBodySocialInteractionsItemInteractionType = typeof CreateDailyLogBodySocialInteractionsItemInteractionType[keyof typeof CreateDailyLogBodySocialInteractionsItemInteractionType];


export const CreateDailyLogBodySocialInteractionsItemInteractionType = {
  FAMILY: 'FAMILY',
  FRIENDS: 'FRIENDS',
  COLLEAGUES: 'COLLEAGUES',
  STRANGERS: 'STRANGERS',
  WORK: 'WORK',
  SHOPPING: 'SHOPPING',
  OTHER: 'OTHER',
} as const;

export type CreateDailyLogBodySocialInteractionsItemSatisfactionLevel = typeof CreateDailyLogBodySocialInteractionsItemSatisfactionLevel[keyof typeof CreateDailyLogBodySocialInteractionsItemSatisfactionLevel];


export const CreateDailyLogBodySocialInteractionsItemSatisfactionLevel = {
  VERY_DISSATISFIED: 'VERY_DISSATISFIED',
  DISSATISFIED: 'DISSATISFIED',
  NEUTRAL: 'NEUTRAL',
  SATISFIED: 'SATISFIED',
  VERY_SATISFIED: 'VERY_SATISFIED',
} as const;

export type CreateDailyLogBodySocialInteractionsItem = {
  interactionType: CreateDailyLogBodySocialInteractionsItemInteractionType;
  satisfactionLevel: CreateDailyLogBodySocialInteractionsItemSatisfactionLevel;
  /** @minimum 0 */
  duration: number;
};

export type CreateDailyLogBodyStressLevelItemStressLevelType = typeof CreateDailyLogBodyStressLevelItemStressLevelType[keyof typeof CreateDailyLogBodyStressLevelItemStressLevelType];


export const CreateDailyLogBodyStressLevelItemStressLevelType = {
  LOW: 'LOW',
  MODERATE: 'MODERATE',
  HIGH: 'HIGH',
  EXTREME: 'EXTREME',
} as const;

export type CreateDailyLogBodyStressLevelItemStressSource = typeof CreateDailyLogBodyStressLevelItemStressSource[keyof typeof CreateDailyLogBodyStressLevelItemStressSource];


export const CreateDailyLogBodyStressLevelItemStressSource = {
  WORK: 'WORK',
  FAMILY: 'FAMILY',
  HEALTH: 'HEALTH',
  FINANCES: 'FINANCES',
  RELATIONSHIPS: 'RELATIONSHIPS',
  OTHER: 'OTHER',
} as const;

export type CreateDailyLogBodyStressLevelItem = {
  stressLevelType: CreateDailyLogBodyStressLevelItemStressLevelType;
  /** @minimum 0 */
  duration: number;
  situationDescription?: string;
  stressSource?: CreateDailyLogBodyStressLevelItemStressSource;
};

export type CreateDailyLogBodySymptomsItemSymptomType = typeof CreateDailyLogBodySymptomsItemSymptomType[keyof typeof CreateDailyLogBodySymptomsItemSymptomType];


export const CreateDailyLogBodySymptomsItemSymptomType = {
  SADNESS: 'SADNESS',
  FATIGUE: 'FATIGUE',
  IRRITABILITY: 'IRRITABILITY',
  LACK_OF_FOCUS: 'LACK_OF_FOCUS',
  ANXIETY: 'ANXIETY',
  INSOMNIA: 'INSOMNIA',
  LOSS_OF_INTEREST: 'LOSS_OF_INTEREST',
  RESTLESSNESS: 'RESTLESSNESS',
} as const;

export type CreateDailyLogBodySymptomsItemSymptomSeverity = typeof CreateDailyLogBodySymptomsItemSymptomSeverity[keyof typeof CreateDailyLogBodySymptomsItemSymptomSeverity];


export const CreateDailyLogBodySymptomsItemSymptomSeverity = {
  MILD: 'MILD',
  MODERATE: 'MODERATE',
  SEVERE: 'SEVERE',
} as const;